requirement_back_env.txt
django 
django-cors-headers
djangorestframework
djangorestframework-simplejwt
pillow
requests
django-environ
psycopg2-binary
python-decouple